tests: Work around race condition in selinuxpolicy status by retrying the is-installed/is-removed check
The selinuxpolicy API object is a singleton and has a single finalizer
and a single status. This means that the first instance of selinuxd that
removes the finalizer or sets the status 'wins' as far as anyone waiting
for status is concerned, but there's no guarantee that the other pods in
the DS running on the other nodes have finished their work as well.

To work around that until we address the issue systematically, let's
retry the check on the nodes up to three times with a sleep of 5 seconds
between the tries.